Fix update endpoint
