+++ api/src/routes/dashboard/auth.rs
@@ -1,13 +1,8 @@
 use std::sync::Arc;
 
-use axum::extract::FromRequestParts;
-use axum::headers::authorization::Bearer;
-use axum::headers::Authorization;
-use axum::http::request::Parts;
-use axum::{async_trait, RequestPartsExt, TypedHeader};
 use axum::{extract::State, Json};
 use base64::{engine::general_purpose, Engine};
-use jsonwebtoken::{decode, encode, Header, Validation};
+use jsonwebtoken::{encode, Header};
 use once_cell::sync::Lazy;
 use sha2::{Digest, Sha512};
 
@@ -15,29 +10,7 @@ use crate::models::auth::{AuthBody, AuthModel, Claims, Keys, LoginPayload};
 use crate::models::error::ApiError;
 use crate::AppState;
 
-#[async_trait]
-impl<S> FromRequestParts<S> for Claims
-where
-    S: Send + Sync,
-{
-    type Rejection = ApiError;
-
-    async fn from_request_parts(parts: &mut Parts, _state: &S) -> Result<Self, Self::Rejection> {
-        // Extract the token from the authorization header
-        let TypedHeader(Authorization(bearer)) = parts
-            .extract::<TypedHeader<Authorization<Bearer>>>()
-            .await
-            .map_err(|_| ApiError::AuthenticationError)?;
-
-        // Decode the user data
-        let token_data = decode::<Claims>(bearer.token(), &KEYS.decoding, &Validation::default())
-            .map_err(|_| ApiError::AuthenticationError)?;
-
-        Ok(token_data.claims)
-    }
-}
-
-static KEYS: Lazy<Keys> = Lazy::new(|| {
+pub static KEYS: Lazy<Keys> = Lazy::new(|| {
     let secret = std::env::var("SECRET").expect("JWT_SECRET must be set");
     Keys::new(secret.as_bytes())
 });
